requirements.txt
pybit
pandas
pandas_ta
python-dotenv # Optional, if using .env file